requirement.txt
psycopg2
Django
selenium
djangorestframework
pygments
django-rest-swagger
drf-yasg
requests